Allow broken Playbacks
In case the Server is not able to lookup the played track and release
group, it will persist a "broken playback" which retains the information
originally send by the client but does not connect the playback to a
recording and release group.

Those broken playbacks can later be fixed by the user explicitly.