Rec: Provide API to accept invites and register a User
A POST request can be used to register a new user using an invite
provided by another User. One can only interact with the recorder
service, if there is a user object in the security context. That is why
we introduce a "fake user" for accepting invites (because there is not a
user yet).

An invited client can login as user $invite$ and use the invite ID as
password. A preliminary user is created and persisted immediately. The
user can then perform a POST request on the register endpoint and set
both a username and a password. Once this is successfully done, the
invite will be invalidated and cannot be used to login anymore. In the
same transaction, the user becomes valid and regular login with the set
password is possible. That way, the invite login is possible until a
real username and password is set.