Add "create user" endpoint
This patch starts the new code base in Kotlin. Apparently, Kotlin is
just as shit as any other language on this planet but you have to write
less! Less shit → better.

Functionally, we introduce a way to create a user, which will
subsequently be protected by security checks, once spring security is
introduced properly again.