Make entity mutable
Hibernate cannot really handle immutable data types due to the way it
instantiates objects when loading them from the database. Thus, we will
not use data classes anymore.
This also implies, that we have to deal with potential null values
within entities. The reason is, that hibernate will use the default
constructor and subsequently set the fields to their respective values.
But then we cannot use val and non-nullable fields anymore.