Add GET endpoint for single user
On the path /users/current or /users/{uuid} one can now retrieve the
details about a single user. The special value current refers to the
currently authenticated user (the client's user itself).

This patch also introduces a application specific Exception
"NotFoundException". It shall be used whenever the client operation is
supposed to retrieve an element but the element was not found. I.e. the
primary object of interest is not present, as opposed to some element
that was needed for a larger operation was not found.
Examples:
- retrieve a user/anything but was not found -> NotFoundException
- add an element referencing a user, but the user was not found -> this
  should *not* result in a NotFoundException since the missing element
  was not the primary object of interest. For this case there should be
  something like a UnresolvedReferenceException or so.

Lastly, we introduce a SecurityHelper which contains a convenience
function to retrieve the currently authenticated User.