Move PlaybackTrans to dedicated package
The translation functions were located in the API package so far, mainly
because the first translation functions that we wrote were used to
translate entities into REST resources.

However, in other packages we already started to use more kinds of
translation functions. Generally, we have two possibilities: have
several translation function types all in one package or distribute the
translation functions to the layers they belong, i.e. translation
functions on api, boundary and data layer.

As it may sometimes not be entirely clear to which layer a translation
function belongs, and they do not seem to be overly complicated, we will
collect them in a dedicated cross-layer package called "trans" for now.
Of this turns out to be chaotic, we will refactor this in the future.