Track musicbrainz relations if detected correctly
Using mbs to identify recordings and release groups, the recorder
service now stores the references to musicbrainz entities in each
Playback. In case this information is available, it will provide the
musicbrainz properties on the API, e.g. for recording title or release
title. If the musicbrainz entry could not be detected, the information
provided by the client will be presented on the API.

A "broken" flag in each playback resource indicates whether the
musicbrainz relation was detected or not.