Add endpoint to update existing playbacks
Introduce a new endpoint PATCH /rec/v1/playbacks/{uuid} to change
existing playbacks. The body contains a resource with fields that need
changing. Since it is PATCH, omitted fields are left unchanged. Only the
owner of a playback can change it.

Externally (API) the resource is identical to the one sent in create
playback. Internally, we introduce a dedicated class, because in the
PATCH case, any field could be null. This is not allowed in the POST
case.

An optional query parameter skipMbs can be set to true, if the client
wishes not to discover the playback on mbs again. This is useful if only
recording-unrelated details were changed.

This patch also introduces the NotAuthorizedException which can be used
to report refusal to perform an operation because the user's rights are
not sufficient.