Add primitive CSV import
This patch introduces a primitive CSV import that enables the user to
migrate last.fm accounts to cfm. The importer can be seen as a first
draft and is not final; among other things, because it is slow as hell.

The main import routine reads a csv file line-by-line and parses it with
predefined patterns. Currently, there is one pattern which matches the
CSV file provided by https://benjaminbenben.com/lastfm-to-csv/.

The importer reports progress and error counters on the command line.
Furthermore, the user can request a fail log with the --fail-log option
which contains all the lines that could not be imported.

Lastly, the recorder service was adapted, because before, we assumed
that all musicbrainz recordings have a non-null length field. This is
not the case. The recorder can now deal with null values.

Future work will speedup importing and make it more resilient.