Decouple PlaybackRes from entity
This preparatory patch cuts the direct link between the PlaybackRes and
the underlying Playback entity. The reason is the change of information
source. PlaybackRes is the user view on a playback. Currently, we store
the whole transitive net of objects necessary to produce this user view.
I.e. the recorder explicitly holds copies of all recordings, release
groups and artists.
In the future, we want to remove these redundant objects. However, in
order to produce the user view on playbacks, we will then need to
fill in the names of recordings, artists and release groups. The
recorder will only hold references (UUIDs) to these objects.

Ideally, this view data inclusion is not done on the entities.
Consequently, we decouple the boundary from the API layer by introducing
a DTO.