Retrieve view data from MBS instead of local entities
Previously, we provided the view data from our local data store, which
keeps a replication of the mbs entities. This patch changes the
behaviour: instead of assuming that all entities are locally available,
we query mbs with a list of UUIDs (of recordings and release groups
respectively) and use the result to provide a view to the user.