Make Playback.toDto() unaware of replicated entities
This patch makes the Playback.toDto() method unaware of the replicated
musicbrainz entities. After the refactoring in the previous patches,
this method will only be called, if a musicbrainz-provided view of the
playback is not available. Consequently, we will only use toDto() to
show original playback data, but nothing managed by musicbrainz.

The replicated entities will be deleted in the following patch(es?) and
thus, this method should work independently.