Recorder can track source information for playbacks
This commit extends the recorder service to track sources for an
individual playback. The source is just a string value.  That way
clients can specify different sources, for example for different devices
or players.

Furthermore, we can use the source field to track imported data
specially. This will be useful later, if we want to do incremental
imports and skip existing data.