Add endpoint to set currently playing track
This commit adds an endpoint "/now" under the playback endpoint. Clients
can PUT playbacks on this endpoint and by that declare a currently
playing track. Naturally, there can only be one such track per user, so
PUTting again will overwrite the previous playback.

Putting playbacks on this endpoint does *not* create a playback as a
POST on the /playback endpoint does. I.e. this is only meant to persist
one single temporary playback to make the currently playing track
retrievable by the user and others.

Because of these properties, the playback is not stored in the regular
playback entity, as this would also display it among the persisted
playbacks.

Note, that the /now endpoint only persists the information we care about
for retrieval again. This is usually a subset of the fields that the
/playback endpoint stores.