Add endpoint to retrieve broken playbacks accumulated
This commit adds a new endpoint under .../playbacks/acc. The endpoint
can be used to retrieve broken playbacks in an accumulated fashion. That
is, an identical playback (same artists/title/release) that occurs n
times makes only *one* entry in the result set, whose "occurrences"
field is n. Moreover, the result set is sorted by occurrences. This
endpoint might be quite useful to retrieve broken playbacks that are
useful to fix.

Accordingly, we might offer an endpoint to "bulk-fix" such identical
playbacks soon.