Sort playbacks with attempted fixes to the end
When a playback cannot be fixed, the client can choose to "not fix" it
for now, which effectively means that we want to sort it to the end of
the list. That is done by the timestamp in fixAttempt. The previous
patch makes sure that fixAttempt is set to a meaningful value, while
thise patch does the actual sorting.