rec: Do not discover mb ids during batch import
Up until now we did "live" discovery of imported playbacks against mbs.
However, since mbs lookup is expensive and we want to implement a
background healing process anyways, we remove the discovery step
entirely. This implies that all imported tracks are "broken" initially.
The naming for "broken" and the possible states (true/false currently)
might be subject to change as well.