rec: Fetch bunches of unattached playbacks
We introduce a new method getUnattachedPlaybackGroups on PlaybackBoundary,
which returns unattached (broken) playbacks from the database in a
grouped manner. Only identical playbacks (same artist, recording,
release, length) are grouped together, since these properties might
influence discovery in musicbrainz.

New names are intentionally introduced ("group" instead of "accumulated"
and "unattached" instead of "broken"), because their behaviour is
different on the API and we want to get rid of the old ones step by
step.

The caller can specify the maximum amount of groups returned. For now,
the API is as simple as that. We assume that the caller wants the
following behaviours as default:

* Groups that have not recently had a fix attempt are reported before
  ones that had. This behaviour makes sure, that all groups are
  eventually tried to be fixed again, since - at some point - all other
  groups have had another attempt.
* Groups with many playbacks are reported before groups with less
  playbacks. This behaviour implies that larger groups with same recency
  of fix attempts are preferred over smaller ones.

These two properties are considered sane non-configurable defaults but
should probably not be relied on for any other interest.

With this patch we also start to use plain JDBC, as we want to move away
from nasty old Hibernate.