rec: Put usual columns in a constant
In PlaybackRepo, playbackView() constructs a list of Playback POJOs from
a result set. This logic is repeated in already two methods and there
are more to follow. Because the playbackView() mapper expects a certain
set of columns with a specific name, we also put the selected columns in
a string constant so we don't have to duplicate it all the time.

This is the downside that it reduces readability of the queries because
I cannot directly see which columns are selected. But at the moment,
duplicated code is the worse problem.